refactor: extract CityBikeStopNearYou query in wrapping container
