+++ app/component/CityBikeStopNearYou.js
@@ -2,7 +2,6 @@ import React, { useEffect } from 'react';
 import PropTypes from 'prop-types';
 import { FormattedMessage } from 'react-intl';
 import { Link } from 'found';
-import { graphql, createRefetchContainer } from 'react-relay';
 import CityBikeStopContent from './CityBikeStopContent';
 import FavouriteBikeRentalStationContainer from './FavouriteBikeRentalStationContainer';
 import { PREFIX_BIKESTATIONS } from '../util/path';
@@ -61,6 +60,7 @@ const CityBikeStopNearYou = ({ stop, relay, currentTime, currentMode }) => {
     </span>
   );
 };
+
 CityBikeStopNearYou.propTypes = {
   stop: PropTypes.object.isRequired,
   currentTime: PropTypes.number.isRequired,
@@ -68,28 +68,4 @@ CityBikeStopNearYou.propTypes = {
   relay: PropTypes.any,
 };
 
-const containerComponent = createRefetchContainer(
-  CityBikeStopNearYou,
-  {
-    stop: graphql`
-      fragment CityBikeStopNearYou_stop on BikeRentalStation {
-        stationId
-        name
-        bikesAvailable
-        spacesAvailable
-        capacity
-        networks
-        state
-      }
-    `,
-  },
-  graphql`
-    query CityBikeStopNearYouRefetchQuery($stopId: String!) {
-      bikeRentalStation(id: $stopId) {
-        ...CityBikeStopNearYou_stop
-      }
-    }
-  `,
-);
-
-export { containerComponent as default, CityBikeStopNearYou as Component };
+export default CityBikeStopNearYou;
